# artifacts/api-server/src/models/User.js
import mongoose from "mongoose";
import bcrypt from "bcryptjs";
const userSchema = new mongoose.Schema({
    name: { type: String, required: true },
    email: { type: String, required: true, unique: true },
    password: { type: String, required: true },
    phone: { type: String, default: null },
    address: { type: String, default: null },
    role: { type: String, enum: ["user", "admin"], default: "user" },
    coins: { type: Number, default: 0 },
    avatar: { type: String, default: null },
}, { timestamps: true });
userSchema.pre("save", async function () {
    if (this.isModified("password")) {
        this.password = await bcrypt.hash(this.password, 10);
    }
});
userSchema.methods.comparePassword = async function (candidate) {
    return bcrypt.compare(candidate, this.password);
};
export const User = mongoose.model("User", userSchema);