# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FoodRush - Food Delivery App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get user cart
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      responses:
        "200":
          description: Cart cleared
  /cart/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponBody"
      responses:
        "200":
          description: Coupon valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "400":
          description: Invalid coupon
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List all available coupons
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "201":
          description: Coupon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
...
  /admin/restaurants:
    post:
      operationId: createRestaurant
      tags: [admin]
      summary: Create restaurant (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRestaurantBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
  /admin/restaurants/{id}:
    put:
      operationId: updateRestaurant
      tags: [admin]
      summary: Update restaurant (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRestaurantBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
    delete:
      operationId: deleteRestaurant
      tags: [admin]
      summary: Delete restaurant (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        address:
          type: string
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        role:
          type: string
        coins:
          type: number
        avatar:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - coins
        - createdAt
    Restaurant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        cuisine:
          type: string
        image:
          type: ["string", "null"]
        rating:
          type: number
        totalRatings:
          type: number
        deliveryTime:
          type: string
        deliveryFee:
          type: number
        minOrder:
          type: number
        address:
          type: string
        isOpen:
          type: boolean
        featured:
          type: boolean
        tags:
          type: array
          items:
            type: string
        menu:
          type: array
          items:
            $ref: "#/components/schemas/MenuItem"
        createdAt:
          type: string
      required:
        - id
        - name
        - cuisine
        - rating
        - totalRatings
        - deliveryTime
        - deliveryFee
        - minOrder
        - address
        - isOpen
        - featured
        - tags
        - createdAt
    MenuItem:
      type: object
      properties:
        id:
          type: string
        restaurantId:
          type: string
        restaurantName:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: string
        image:
          type: ["string", "null"]
        rating:
          type: number
        totalRatings:
          type: number
        isVeg:
          type: boolean
        isAvailable:
          type: boolean
        isBestseller:
          type: boolean
        tags:
          type: array
          items:
            type: string
        suggestedWith:
          type: array
          items:
            type: string
      required:
        - id
        - restaurantId
        - restaurantName
        - name
        - price
        - category
        - rating
        - totalRatings
        - isVeg
        - isAvailable
        - isBestseller
        - tags
        - suggestedWith
...
    Cart:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        restaurantId:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        couponCode:
          type: ["string", "null"]
        discount:
          type: number
      required:
        - id
        - userId
        - items
        - subtotal
        - deliveryFee
        - total
        - discount
    CartItem:
      type: object
      properties:
        id:
          type: string
        menuItemId:
          type: string
        name:
          type: string
        price:
          type: number
        quantity:
          type: number
        image:
          type: ["string", "null"]
        restaurantId:
          type: string
      required:
        - id
        - menuItemId
        - name
        - price
        - quantity
        - restaurantId
    AddToCartBody:
      type: object
      properties:
        menuItemId:
          type: string
        quantity:
          type: number
      required:
        - menuItemId
        - quantity
    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: number
      required:
        - quantity
...
    Coupon:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
        discountValue:
          type: number
        minOrderAmount:
          type: number
        maxDiscount:
          type: ["number", "null"]
        isActive:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        usageLimit:
          type: ["number", "null"]
        usedCount:
          type: number
      required:
        - id
        - code
        - description
        - discountType
        - discountValue
        - minOrderAmount
        - isActive
        - usedCount
    ValidateCouponBody:
      type: object
      properties:
        code:
          type: string
        orderAmount:
          type: number
      required:
        - code
        - orderAmount
    CreateCouponBody:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
        discountType:
          type: string
        discountValue:
          type: number
        minOrderAmount:
          type: number
        maxDiscount:
          type: number
        usageLimit:
          type: number
        expiresAt:
          type: string
      required:
        - code
        - description
        - discountType
        - discountValue
        - minOrderAmount
...
    CreateRestaurantBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        cuisine:
          type: string
        image:
          type: string
        deliveryTime:
          type: string
        deliveryFee:
          type: number
        minOrder:
          type: number
        address:
          type: string
        isOpen:
          type: boolean
        featured:
          type: boolean
        tags:
          type: array
          items:
            type: string
      required:
        - name
        - cuisine
        - deliveryTime
        - deliveryFee
        - minOrder
        - address
...